requirements.txt
fastapi
uvicorn
python-dotenv
openai
langchain
tiktoken
sqlalchemy
PyPDF2